api/pyproject.toml
[project]
name = "holocron"
version = "0.2.2.dev0"
description = "Backend template for your Vision API with Holocron"
requires-python = ">=3.11,<4.0"
license = { text = "Apache-2.0" }
authors = [{ name = "François-Guillaume Fernandez", email = "fg-feedback@protonmail.com" }]
maintainers = [{ name = "François-Guillaume Fernandez", email = "fg-feedback@protonmail.com" }]
readme = "README.md"
keywords = ["backend", "api", "computer vision", "fastapi", "onnx"]
dependencies = [
    "fastapi>=0.116.1,<1.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "uvicorn>=0.23.0,<1.0.0",
    "python-multipart>=0.0.20,<1.0.0",
    "Pillow>=8.4.0,!=9.2.0",
    "onnxruntime>=1.22.0,<2.0.0",
    "huggingface-hub>=1.0.0,<2.0.0",
    "numpy>=1.19.5,<3.0.0"
]

[project.optional-dependencies]
test = [
    "pytest==8.4.2",
    "pytest-asyncio>=1.1.0,<2.0.0",
    "httpx>=0.28.1,<1.0.0",
    "pytest-cov==6.2.1",
    "pytest-pretty==1.3.0",
    "requests>=2.32.0,<3.0.0"
]
